Split test_salpy.py into two files
to avoid a segfault caused by creating a SALPY object and a dds Domain in the same process.
